Extract loading and error functions
