+++ media/main.js
@@ -4,6 +4,9 @@ while (!$) {
 import replaceIssueLinks from "./utils/replaceIssueLinks.js";
 import replaceUserTags from "./utils/replaceUserTags.js";
 import dateToHumanReadable from "./utils/dateToHumanReadable.js";
+import setReceivedError from "./utils/setReceivedError.js";
+import setLoading from "./utils/setLoading.js";
+import removeLoading from "./utils/removeLoading.js";
 const vscode = acquireVsCodeApi();
 
 const link = document.getElementsByClassName("create-docs");
@@ -120,80 +123,19 @@ $(document).ready(function () {
       });
     });
   };
-  function setLoading() {
-    $("#ghHolder").replaceWith(`
-    <div id="ghHolder">
-      <p>Loading...</p>
-    </div>
-    `);
-    errorTimeout = setTimeout(setError, 4000);
-  }
-  function setError() {
-    $("#ghHolder").replaceWith(`
-    <div id="ghHolder">
-      <p>We might have run into an error, our team is on it.</p>
-      <p>Try running a new Watermelon query, please.</p>
-    </div>
-    `);
-    $("#ghHolder").append(
-      "<button class='run-watermelon'>Run Watermelon</button><br/>"
-    );
-    $(".run-watermelon").on("click", (event) => {
-      sendMessage({ command: "run" });
-    });
-    $("#ghHolder").append(
-      `<button class='create-docs' >Create repo docs</button>`
-    );
-    $(".create-docs").on("click", (event) => {
-      sendMessage({ command: "create-docs" });
-    });
-
-    $("#ghHolder").append(
-      "<p>Alternatively, you can <a href='https://github.com/watermelontools/wm-extension#commands'>run with our watermelon.start command</a></p>"
-    );
-    $("#ghHolder").append(
-      "<p>Select a piece of code to start. Then run the Watermelon VS Code Command by pressing <kbd>CTRL</kbd> + <kbd>SHIFT</kbd> + <kbd>P</kbd> (or <kbd>CMD</kbd> + <kbd>SHIFT</kbd> + <kbd>P</kbd> in Mac) and type > <code>start watermelon</code></p>"
-    );
-  }
-  function setReceivedError(errorText) {
-    clearTimeout(errorTimeout);
-    $("#ghHolder").replaceWith(`
-    <div id="ghHolder">
-      <p>We ran into this error: ${errorText}</p>
-      <p>Try running a new Watermelon query, please.</p>
-    </div>
-    `);
-    $("#ghHolder").append(
-      "<button class='run-watermelon'>Run Watermelon</button><br/>"
-    );
-    $(".run-watermelon").on("click", (event) => {
-      sendMessage({ command: "run" });
-    });
-    $("#ghHolder").append(
-      `<button class='create-docs' >Create repo docs</button>`
-    );
-    $(".create-docs").on("click", (event) => {
-      sendMessage({ command: "create-docs" });
-    });
-  }
-  function removeLoading() {
-    clearTimeout(errorTimeout);
-    $("#ghHolder p").remove();
-    $("#ghHolder button").remove();
-  }
 
   window.addEventListener("message", (event) => {
     const message = event.data; // The JSON data our extension sent
     switch (message.command) {
       case "prs":
-        removeLoading();
+        removeLoading(errorTimeout);
         addPRsToDoc(message.data, event.data.explanation);
         break;
       case "loading":
         setLoading();
         break;
       case "error":
-        setReceivedError(message.error.errorText);
+        setReceivedError(message.error.errorText, errorTimeout);
         break;
       case "author":
         authorName = message.author;
