+++ media/utils/removeLoading.js
@@ -1,0 +1,6 @@
+function removeLoading(errorTimeout) {
+    clearTimeout(errorTimeout);
+    $("#ghHolder p").remove();
+    $("#ghHolder button").remove();
+  }
+  export default removeLoading;
